API/Reference

API/Reference

Index

SkyCoords.AbstractSkyCoords

SkyCoords.FK5Coords

SkyCoords.GalCoords

SkyCoords.ICRSCoords

SkyCoords.separation

Types

SkyCoords.AbstractSkyCoords — Type.

The supertype for all sky coordinate systems.

source

SkyCoords.ICRSCoords — Type.

ICRSCoords(ra, dec)

International Celestial Reference System

This is the current standard adopted by the International Astronomical Union notably due to its high level of accuracy compared to standard equatorial coordinate systems. What sets this apart from FK5Coords is that it is completely defined using extragalactic radio sources rather than a geocentric frame, which means the reference frame will not change due to Earth's motion.

Coordinates

ra - Right ascension in radians (0, 2π)

dec - Declination in radians (-π/2, π/2)

source

SkyCoords.GalCoords — Type.

GalCoords(l, b)

Galactic Coordinate System

This coordinate system is defined based on the projection of the Milky Way galaxy onto our celestial sphere, with (0, 0) being approximately the center of our galaxy.

Coordinates

l - Galactic longitude in radians (-π, π)

b - Galactic latitude in radians (-π/2, π/2)

source

SkyCoords.FK5Coords — Type.

FK5Coords{equinox}(ra, dec)

Equatorial Coordinate System

This coordinate system maps the celestial sphere based on a geocentric observer. Historically the oldest, this coordinate system has been shown to be inaccurate due to its definitions based on the Earth, which has long-scale precession causing the reference frame to change. Because of this, an equinox must be provided (typically 2000, commonly known as J2000) which defines the reference frame.

Coordinates

ra - Right ascension in radians (0, 2π)

dec - Declination in radians (-π/2, π/2)

source

Functions

SkyCoords.separation — Function.

separation(c1::AbstractSkyCoords, c2::AbstractSkyCoords) -> distance

Return angular separation between two sky coordinates, in radians.

The angular separation is calculated using the Vincenty formula, which is slightly more complex and computationally expensive than some alternatives, but is stable at at all distances, including the poles and antipodes.

source